Updated README, Example Graph Illustration
Filled README with text about the goal, as well as technical notes. In
addition to this, I created an illustration of a simple graph that this
program will be able to decipher. It is used as an example in the README
to demonstrate how the graphs will be represented in the program.