Laptop source code commit
Now that visual studio is working, I want to test a slightly modified
source .cpp file, just to make sure everything is working fine.

I get some kind of errors, or files not found, that I believe are
because I'm using the 2013 version of studio. If this continues as an
issue I'll just clear my repository of the 2012 files and replace them
with the 2013.