Text Interface Implementation
Goal: Create a basic text based interface for a user to select which
map(s) they want to work with

Accomplished: Skeleton for the interface. Menu loops properly and
displays introductory text as well as options for the user. These
options loop and call the appropriate functions based on user input. The
functions called only have "filler" content, but will be ready to
fulfill their respective duties once the Map class is properly made.

Next: Begin writing the foundations for the Map class, as this is the
center of the project. This will begin by writing the Node class, as a
Map is a network of Nodes and a collection of other data (eg number of
nodes, names, possibly an AllPaths class, not sure yet).

Other thoughts: For now a text based menu is fine, but I'd like to
eventually come back to this project once the "meat" of this program is
finished. Instead of text based, I'd like to make a GUI, where you load
your map using mouse clicks instead of typing anything. Additionally I'd
like to add a feature where a user can build a map, rather than having
to write a tedious map file with very specific syntax. However, that's
almost an entirely separate project. The point of this one is to work
with the shortest path algorithm, not working with a GUI.

However, with that being said, I feel like doing this project with OOP
designs makes that upgrade significantly easier. The Map API won't
change just because the format of the way it's being accessed is. When I
move on to that, I won't have to deal with any of the actual shortest
path stuff I want to focus on here. In my old way of trying to do this,
in C, that was impossible.