Node Class Implementation In Progress 1
Goal: Fully implement the Node class, to be ready to be used by Map
class

Accomplished: Began working on Node class. Made a decision that, in
order to have a given node be able to "point" to all other nodes it
connected to, that I would make a Path object, and a PathLinkedList
class. A Path object's main fields are Node A and B, and a path weight.
PathLinkedList will be just what it sounds like - a linked list made of
these Paths. In order to be able to make a linked list out of Path
objects, the Path class has the additional fields lastPath and nextPath,
to point to the last/next Path in the list appropriately.

PathLinkedList is currently incomplete. I didn't make the pointer fields
in Path mutable, so I need to go back and change that. Once that's done,
I will finish writing the append function for PathLinkedList. After
that, I will overload the '+' operator so that I can append a Path to a
PathLinkedList using just that operator.

Next: Once PathLinkedList is finished and working, the Node class will
most be done. This is the only really tricky part, since this will
contain all of the information needed. The difficult part will be what
comes after that - the map file text parsing. However, I'm not sure at
this moment exactly when I want to start dealing with that. I could
possibly do it immediately after PathLinkedList is working, but the
testing would involve altering the fields to be mutable. The alternative
would be to do the text parsing after I've written the Map class, but
that's a lot of code I'll have written without testing. It's yet to be
determined.