Debugging Statements for PathLinkedList
Not done working on the project for the day just yet, but I finished
PathLinkedList, and decided to start debugging it to make sure it worked
before I started using it in the Node class. It currently works, and is
FULL of debugging statements. I don't want to keep them in the code,
even as comments, but I want to record them so I'm taking a snapshot.

After cleaning up my code, I'm going to start integrating PathLinkedList
with the Node class.

The biggest issue I had was remember that (apparently) classes aren't
passed by reference by default. This was causing issues with pointers
everywhere. Once I realized (ie remembered) that I needed to be
explicit, everything fell together! Need to be more careful about these
things, but I guess that comes with shaking the rust off my C++