Node Debugging Statements and Flaw in Path Class
Pushing my debugging statements in main to Git to keep track of them.
Additionally, I realized I made a HUGE error in the design of my
PathLinkedList class.

Each Path has a forward and back pointer to the next Path in the
PathLinkedList. This would be fine if each Path were in ONLY one
PathLinkedList - but they're not. So if a Path is added to another
PathLinkedList, its forward/back pointers will be overwritten, basically
destroying the linked lists in all PathLinkedLists that path has been
added to before.

I have to re-design this. Additionally, my friend pointed out what I've
been calling Paths are really Edges. I need to go back and change
everything I've been calling a Path an Edge, to clarify. Might as well
do that.