Fixing Edge Linked List
I'm currently in the process of fixing the problem stated in the
previous push - that the pointers used to make the linked list are
members of the Edge objects, which becomes a problem when there a single
edge is in multiple link lists. I'm changing it so that the pointers are
in the EdgeLinkedList objects instead.

This is proving troublesome. I've made a struct called EdgeCell, which
contains 3 fields - a data field (of type Edge), a back pointer and a
forward pointer. This means I have to change how the append and getIndex
functions work. It's pointers galore.

It sort of works. I can get through some of the original debugging
statements I had before. However, once it starts trying to iterate
through all the Edges connected to a given Node, I get segfaults after
the first Edge's info is displayed.

Whether or not the problem is in getIndex, append, or both functions, I
don't know. I'm going to dig into it next time. Getting this working
will be the next step.