Creating Graph Class - Determine number of Nodes
I've begun creating the Graph class. I've decided it will be immutable,
and it will be created only by calling the constructor with the filename
of the corresponding graph file as a parameter.

The design philosophy is inefficient, I know, but I want to make it look
"neat", and abstract. Because of this, the constructor itself only
opens/closes the file once (without doing anything to it), to make sure
the file actually exists. If it doesn't, it will throw an error (I will
decide exactly how to do this when I write the LoadMap function in
TextUI.

The constructor will use helper methods. The first such method is going
to be findNumOfNodesInFile(string). It does exactly what it sounds like
- given a graph file, it determines the number of nodes in that graph.
I've included the debugging statements I used to determine that it
works.

The next step is to get the nodeList properly allocated. I do believe
this will require pointer tricks, because I have to give it an initial
amount when I declare it in the header file. I think I can do something
like "nodeList = new Node[numOfNodes];" or something like that. We'll
see.

If that doesn't work, I'll make another linked list - although I won't
make a completely separate class. I'll just make a struct in the Graph
class to manage everything.