Creating Graph Class - Initializing Array of Nodes
So my idea of making an array from a pointer worked, though it's a
little different than what I was exactly looking at. The statement was
actually "nodeList = new Node[numOfNodes];", where nodeList was declared
in the header as so: "Node *nodeList;". Works perfectly. I kept the
debugging statements in this push (again) to show that.

The next step is going to be to actually parse each individual line in
the graph file to populate the Nodes with the correct information (eg
Node 1 connects to Node 5 via an Edge of weight 7, etc...). Not sure how
I'm going to collect this data (Using C-strings, like I have in the
past, Regexes, or something else? Need to do research). Additionally,
I'm not sure off the top of my head how I'm going to allocate this
information, though that seems like it'll be the easier of these two
issues.