Creating Graph Class - Graph File Parsing
I decided to go with using regexes to parse the text in the Graph files.
This is because I had forgotten, when I did this with C-strings, it was
going to be a HUGE pain to have anything with 10 or more Nodes or path
weights greater than 10. Sure I COULD do it, but regexes (when they
work) look sexier, and I need to get over my fear of them.

It took some study, but I'm glad I went in this direction. The code I
used (as far as the regexes are concerned) are pretty neat and tidy.
They do work, and I can now successfully grab the "Node nuggets", as
I've taken to calling them, with as many digits as I please.

To discuss how I'm doing this overall, the constructor calls the method
parseGraphFile, which in turn calls parseGraphLine for, you guessed it,
every line (that defines a Node) in that File. parseGraphLine is
actually going to be recursive. I will explain how using an example.

Given a line with the 3 Node nuggets:
"1, 10 2,20 3,30"
parseGraphFile will pass the (full) line "1,10 2,20 3,30" to
parseGraphLine. parseGraphLine will then find the first and ONLY the
first Nugget - "1,10", and do everything that needs to be done (eg make
or re-use the Edge that connects this Node with Node 1). It will then
truncate this Nugget (and the following space character) from the line,
making the line now "2,20 3,30". parseGraphLine will then call itself
using THIS line, until the line no longer contains any Nuggets. Once the
line is depleted, it will go back up the stack, and parseGraphFile will
then move on to the next line in the file.

This is not fully implemented yet. I have comments outlining how I'd
like to do this, and I have to check to see if the line is depleted -
however, I'm not sure it's where I want or need it to be yet.
Additionally I haven't implemented the actual recursive part, as I
haven't written the code for truncating the first Node nugget from the
line.

Currently I'm on the part where I make a new Edge or determine if one is
in use. I've written a Node method (doNodesConnect) to determine if two
Nodes do in fact connect. This method returns a pointer to the Edge that
connects these two Nodes, or NULL if none exists.

It doesn't work, and I'm in the middle of debugging it, so I'm leaving
all my debugging statements in for this push. Additionally, it only runs
the check based on that Node's nodeNumber. Now that I think about it,
I'm going to check by the address of the Node to be sure.