Creating Graph Class - Debugging doNodesConnect
What a headache. I have 10 million debugging statements that all look
horribly complicated. I haven't made any real progress since the last
push, but I've been working on it. Just a big case of the stucks.

Specifically, as mentioned, I have parseGraphFile, which calls
parseGraphLine for every line in the File (obviously). The problem is
that each time parseGraphLine is being called, pointers are being
changed in things that they shouldn't be.

Even more specifically, in each instance of parseGraphLine where a new
Edge between two Nodes is discovered, a new Edge is made, and that same
Edge is put into each of the two Nodes' connectingEdges object. The Edge
object is being placed into each linked list correctly, and it is
unique. The problem is that between iterations, the Edge's pointers to
Node A and Node B are being changed to the restricted memory address
0xCCCCCCCC. It is ALWAYS this same address. I have no idea why.

I've spent essentially all of today discovering this, and narrowing it
down to WHERE. This is not occurring in the append(), getEdge(),
convertEdgeToEdgeCell(), operator+(), or addEdge() methods. It is for
SURE happening between iterations of parseGraphLine().

I have NO idea why. Everything is pointers, nothing should be changing.
I'll try to get this figured out.