Creating Graph Class - Fixed doNodesConnect
I got it working. My friend told me it sounded like some of my objects
were being deleted as they were changing scope - this causes illegal
memory address assignment. My program was changing scope - calls to
parseGraphLine other than the original one in which it was created would
cause my Edge objects to behave strangely.

However, this lead me to believe the problem was in the Node object,
because those were the ones being assigned illegal memory addresses.
This was not the case. Turns out the problem was in my Edge class.
Originally, I was using just the constructor to create my Edge objects,
but apparently I needed to be more organized. Using the new keyword
fixed this issue, and I'm able to proceed. I'm leaving in some of the
debugging statements to show that it is in fact working.