Creating Graph Class - Completed File Parsing
Since I was (apparently) so thorough getting things working with this
Edge instantiation issue, all of my Node Linked List stuff works, and I
just had to tweak some regular expressions to get everything working. It
now completely and properly allocates all of the Nodes and Edges
declared in a Graph text file. I left the debugging statements and
temporary friend classing in there for this push.

The next step is to return to the interface, and give the user the
option to load their Graphs into the program, and all those other
necessary options.