Interfacing Graph Class - Loading Maps
I'm in the middle of writing the loadMap method. Not stuck, just have to
go to work. I have it throwing an error if the Graph constructor can't
find/load the file specified by the user.

I need to go back and change everything I had that was "map" to "graph".
The next step as far as programming goes is that I want to write my own
exception class, because I want Graph to be able to throw multiple kinds
of them, and have them caught in different ways.