Interfacing Graph Class - GFNF Exception, name
Created a dedicated class for the GraphFileNotFound exception. Fairly
straightforward, didn't bother making an implementation file, as there's
only one function (overloaded what())with literally 1 line of code in
it. It is thrown when, surprise, the user enters a Graph file name that
can't be found.

Additionally, I added a name field to the Graph class. It's set in the
constructor, based on the string passed to it. I use a regex to delete
the file name. I included the debugging statements I used to show that
it works. The GFNF also works.