Interfacing Graph Class - Fixing Graphs Not Being Saved Issue
I worry about myself sometimes. The reason the Graphs weren't being
saved in the Collection was because I was instantiating the
GraphCollection object inside the loop... So it was being
re-instantiated every iteration. Woops.

I also tweaked the double-entry filter, and that works perfectly now.

Next I'm going to tidy things up. Currently you get an infinite loop if
the user enters a selection that isn't an integer. I'm going to fix that
using regular expressions. Additionally, if the user wants to exit the
program, the program doesn't delete/manage any of the memory it
allocated, so I'm going to write/call all of the destructors
appropriately.