Interfacing Graph Class - User Input
I successfully implemented the user input check using regular
expressions. It was a little bit of trouble because my searching regex
just grabbed the first char of the user input. However, this is of
course a problem in the case of garbage input such as "11dlsjlidj",
because it would grab 1 and think that was valid. So I added a search to
see if the starting sequence is 2 or more of the same digit twice in a
row. It works just fine now.

I played around with proper memory deletion but the truth is I should
have done it from the beginning. Going back and making sure all the
destructors work at this point is not worth the effort when I want to
focus on the Path class from here out. Additionally, it might actually
be beyond the scope of my education so far. I'm just going to leave it
as is, and let the program end without calling destructors
appropriately. At least I know everything is being allocated
efficiently.