Interfacing Graph Class - Select Graph
Implemented Options for Selecting a Graph. This went very well! However
before that I had to fix my user input verification. It was actually
able to fooled by a single digit, a string of numbers, and another
digit. This was of course an issue, but I actually fixed it. Instead of
using regexes, however, I used isdigit() to make sure the first char in
the string was in fact a digit, and string.length(). to make sure the
string was exactly one character long. Digits in the range of 0, 5-9 are
handled by the nature of the switch-case statements and the while loop.
Works perfectly now.

I went on to write the "SelectGraph" menu. I decided to have 4 options:
1. Find shortest Path
2. Diagnostic Information
3. Delete Graph
4. Go back to Program Options

They're all fairly self explanatory. The only one I've fully implemented
is option 2: diagnostic info. It gives the Graph name, number of Nodes,
and then displays what other Nodes each Node connects to, with that
connecting Edge's weight. It works perfectly. The next step is to write
a method to delete a Graph from a GraphCollection... However with how
messy my destructor may have to be I might scrap that.

The step after that is to FINALLY start writing my Path class to find
the shortest Path. Sweet :)